requirements.txt
Django==3.1.1
django-crispy-forms==1.9.2
django-rest-auth==0.9.5
djangorestframework==3.12.1
django-heroku==0.3.1
gunicorn==20.0.4
httplib2==0.9.2
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
MarkupSafe==1.1
numpy==1.19.1
oauth==1.0.1
python-dateutil==2.8.1
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
sqlparse==0.3.1
urllib3==1.22
django-braces==1.14.0
django-cors-headers==3.5.0
django-allauth==0.41.0
